New tenet and depth first node smelling
Add a new tenet: when asserting within a loop, the length of the
collection being iterated should be asserted.

Make SmellNode finish its ast walk before the next node is started.

Add code to support nested smells. This is not complete. The
new code is commented out, with comments, to come back to.