Fix loop assert tenet bug. Also make slasher unit tests pass.
loop assert tenet was panicking because it tried to get the object of a
var that was declared outside the file. This PR is a quick fix, the
proper fix is to get the ast walk using the a token.FileSet. Possibly
all files of a package will need to be added to the FileSet before any
smell is started.